requirements.txt
Backtesting==0.3.3
bokeh==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.45.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
MetaTrader5==5.0.45
neat-python==0.92
numpy==1.26.2
nx==0.0.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sklearn==0.0
threadpoolctl==3.2.0
tornado==6.4
tzdata==2023.3
xyzservices==2023.10.1